food and user seed add

+++ database/migrations/2017_05_09_174949_create_staticorders_table.php
@@ -31,6 +31,8 @@ class CreateStaticordersTable extends Migration
      */
     public function down()
     {
-        Schema::drop('staticorders');
+        if (Schema::hasTable('staticorders')) {
+            Schema::drop('staticorders');
+        }
     }
 }
