+++ database/migrations/2017_05_11_123453_drop_staticorder.php
@@ -12,7 +12,9 @@ class DropStaticorder extends Migration
      */
     public function up()
     {
-        Schema::drop('staticorders');
+        if (Schema::hasTable('staticorders')) {
+            Schema::drop('staticorders');
+        }
     }
 
     /**
